Added Restrictions to pages

+++ DjangoDiscourse/chat/views.py
@@ -1,6 +1,7 @@
 from django.shortcuts import render,redirect
 from django.contrib.auth.models import User
 from django.db.models import Q
+from django.contrib.auth.decorators import login_required
 from django.contrib.auth import authenticate,login,logout
 from django.contrib import messages
 from .models import *
@@ -36,6 +37,8 @@ def room(request,pk):
     context = {'room' : room}
     return render(request,'chat/room.html',context)
 
+
+@login_required(login_url='loginpage')
 def createRoom(request):
     form = RoomForm()
     if request.method ==  'POST':
@@ -47,9 +50,14 @@ def createRoom(request):
     return render(request,'chat/room_form.html',context)
 
 
+@login_required(login_url='loginpage')
 def updateRoom(request,pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if  request.user != room.host:
+        messages.error(request,"Only Room Owner can edit!")
+        return redirect('home')
 
     if request.method == 'POST':
         form = RoomForm(request.POST,instance=room)
@@ -60,8 +68,15 @@ def updateRoom(request,pk):
     context = {'form' : form}
     return render(request,'chat/room_form.html',context)
 
+
+@login_required(login_url='loginpage')
 def deleteRoom(request,pk):
     room = Room.objects.get(id=pk)
+
+    if  request.user != room.host:
+        messages.error(request,"Only Room Owner can edit!")
+        return redirect('home')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
@@ -70,12 +85,17 @@ def deleteRoom(request,pk):
 
 
 def loginPage(request):
+
+
+    if request.user.is_authenticated:
+        return redirect('home')
+
     if request.method == 'POST':
         username = request.POST.get('username')
         password = request.POST.get('password')
 
         try:
-            user = User.objects.get('username')
+            user = User.objects.get(username = username)
         except: 
             messages.error(request,'User dont exists..')
             return redirect('loginpage')
